# [cann] RoPE operator optimization